+++ src/screens/premium/premium-screen.tsx
@@ -1,7 +1,7 @@
 import { deleteRxDBReplication, useRxdbPremiumStart } from '../../rxdb/use-rxdb-premium-start';
 import { ProductList } from './components/product-list';
 import React from 'react';
-import { Button, Text, View } from 'react-native';
+import { Alert, Button, Text, View } from 'react-native';
 
 export function PremiumScreen() {
   const { isReady } = useRxdbPremiumStart();
@@ -13,9 +13,25 @@ export function PremiumScreen() {
     <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
       <Text>RxDB Premium</Text>
       <Button
-        title={'Delete Database'}
+        title={'Delete RxDB Premium Database'}
         onPress={async () => {
-          await deleteRxDBReplication();
+          Alert.alert(
+            'Delete RxDB Premium Database',
+            'Are you sure you want to delete the RxDB Premium database?',
+            [
+              {
+                text: 'Cancel',
+                style: 'cancel',
+              },
+              {
+                text: 'OK',
+                onPress: async () => {
+                  await deleteRxDBReplication();
+                },
+              },
+            ],
+            { cancelable: false }
+          );
         }}
       />
       <ProductList />
